Separate Schema into Schema and Schemas
