+++ cdap-api/src/main/java/co/cask/cdap/api/data/schema/Schema.java
@@ -17,24 +17,9 @@
 package co.cask.cdap.api.data.schema;
 
 import co.cask.cdap.api.annotation.Beta;
-import co.cask.cdap.internal.io.SQLSchemaParser;
-import co.cask.cdap.internal.io.SchemaTypeAdapter;
-import com.google.common.base.Preconditions;
-import com.google.common.base.Throwables;
-import com.google.common.collect.BiMap;
-import com.google.common.collect.HashMultimap;
-import com.google.common.collect.ImmutableBiMap;
-import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.ImmutableSet;
-import com.google.common.collect.Iterables;
-import com.google.common.collect.Maps;
-import com.google.common.collect.Multimap;
-import com.google.common.io.CharStreams;
-import com.google.gson.stream.JsonWriter;
 
-import java.io.IOException;
-import java.io.Reader;
+import java.util.AbstractMap;
+import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -45,7 +30,6 @@ import java.util.Set;
  */
 @Beta
 public final class Schema {
-  private static final SchemaTypeAdapter SCHEMA_TYPE_ADAPTER = new SchemaTypeAdapter();
 
   /**
    * Types known to Schema.
@@ -124,224 +108,6 @@ public final class Schema {
     }
   }
 
-  /**
-   * Parse the given JSON representation, as returned by {@link #toString()} into a Schema object.
-   *
-   * @param schemaJson the json representation of the schema
-   * @return the json representation parsed into a schema object
-   * @throws IOException if there was an exception parsing the schema
-   */
-  public static Schema parseJson(String schemaJson) throws IOException {
-    return SCHEMA_TYPE_ADAPTER.fromJson(schemaJson);
-  }
-
-  /**
-   * Parse the given json representation of a schema object contained in a reader into a schema object.
-   *
-   * @param reader the reader for reading the json representation of a schema
-   * @return the parsed schema object
-   * @throws IOException if there was an exception parsing the schema
-   */
-  public static Schema parseJson(Reader reader) throws IOException {
-    return SCHEMA_TYPE_ADAPTER.fromJson(reader);
-  }
-
-  /**
-   * Parse the given sql-like schema representation into a schema object. Input is expected to be a comma
-   * separated list of names and types, where the type is a primitive type, array, map, record, or union.
-   * The input will be parsed in a case insensitive manner.
-   *
-   * <pre>
-   * {@code
-   * column_name data_type, ...
-   *
-   * data_type : primitive_type
-   * | array_type
-   * | map_type
-   * | record_type
-   * | union_type
-   * [ not null ]
-   *
-   * primitive_type: boolean
-   * | int
-   * | long
-   * | float
-   * | double
-   * | bytes
-   * | string
-   *
-   * array_type: array<data_type>
-   *
-   * map_type: map<data_type, data_type>
-   *
-   * record_type: record<field_name:data_type, ...>
-   *
-   * union_type: union<data_type, data_type, ...>
-   * }
-   * </pre>
-   *
-   * @param schemaString the schema to parse
-   * @return the parsed schema object
-   * @throws IOException if there was an exception parsing the schema
-   */
-  public static Schema parseSQL(String schemaString) throws IOException {
-    return new SQLSchemaParser(schemaString).parse();
-  }
-
-  /**
-   * Creates a {@link Schema} for the given type. The type given must be a
-   * {@link Schema.Type#isSimpleType() Simple Type}.
-   *
-   * @param type Type of the schema to create.
-   * @return A {@link Schema} with the given type.
-   */
-  public static Schema of(Type type) {
-    Preconditions.checkArgument(type.isSimpleType(), "Type %s is not a simple type.", type);
-    return new Schema(type, null, null, null, null, null, null, null);
-  }
-
-  /**
-   * Creates a nullable {@link Schema} for the given schema, which is a union of the given schema with the null type.
-   * The type given must not be the null type.
-   *
-   * @param schema Schema to union with the null type.
-   * @return A nullable version of the given {@link Schema}.
-   */
-  public static Schema nullableOf(Schema schema) {
-    Preconditions.checkArgument(schema.type != Type.NULL, "Given schema must not be the null type.");
-    return Schema.unionOf(schema, Schema.of(Type.NULL));
-  }
-
-  /**
-   * Creates a {@link Schema} of {@link Type#ENUM ENUM} type, with the given enum values.
-   * The set of values given should be unique and must contains at least one value.
-   * The ordering of values in the enum type schema would be the same as the order being passed in.
-   *
-   * @param values Enum values.
-   * @return A {@link Schema} of {@link Type#ENUM ENUM} type.
-   */
-  public static Schema enumWith(String...values) {
-    return enumWith(ImmutableList.copyOf(values));
-  }
-
-  /**
-   * Creates a {@link Schema} of {@link Type#ENUM ENUM} type, with the given enum values.
-   * The set of values given should be unique and must contains at least one value.
-   * The ordering of values in the enum type schema would be the same as the {@link Iterable#iterator()} order.
-   *
-   * @param values Enum values.
-   * @return A {@link Schema} of {@link Type#ENUM ENUM} type.
-   */
-  public static Schema enumWith(Iterable<String> values) {
-    Set<String> uniqueValues = ImmutableSet.copyOf(values);
-    Preconditions.checkArgument(uniqueValues.size() > 0, "No enum value provided.");
-    Preconditions.checkArgument(Iterables.size(values) == uniqueValues.size(), "Duplicate enum value is not allowed.");
-    return new Schema(Type.ENUM, uniqueValues, null, null, null, null, null, null);
-  }
-
-  /**
-   * Creates a {@link Schema} of {@link Type#ENUM ENUM} type, with values extracted from the given {@link Enum} class.
-   * The ordering of values in the enum type schema would be the same as the {@link Enum#ordinal()} order.
-   *
-   * @param enumClass Enum values.
-   * @return A {@link Schema} of {@link Type#ENUM ENUM} type.
-   */
-  public static Schema enumWith(Class<Enum<?>> enumClass) {
-    Enum<?>[] enumConstants = enumClass.getEnumConstants();
-    String[] names = new String[enumConstants.length];
-    for (int i = 0; i < enumConstants.length; i++) {
-      names[i] = enumConstants[i].name();
-    }
-    return enumWith(names);
-  }
-
-  /**
-   * Creates an {@link Type#ARRAY ARRAY} {@link Schema} of the given component type.
-   * @param componentSchema Schema of the array component.
-   * @return A {@link Schema} of {@link Type#ARRAY ARRAY} type.
-   */
-  public static Schema arrayOf(Schema componentSchema) {
-    return new Schema(Type.ARRAY, null, componentSchema, null, null, null, null, null);
-  }
-
-  /**
-   * Creates a {@link Type#MAP MAP} {@link Schema} of the given key and value types.
-   * @param keySchema Schema of the map key.
-   * @param valueSchema Schema of the map value
-   * @return A {@link Schema} of {@link Type#MAP MAP} type.
-   */
-  public static Schema mapOf(Schema keySchema, Schema valueSchema) {
-    return new Schema(Type.MAP, null, null, keySchema, valueSchema, null, null, null);
-  }
-
-  /**
-   * Creates a {@link Type#RECORD RECORD} {@link Schema} of the given name. The schema created
-   * doesn't carry any record fields, which makes it only useful to be used as a component schema
-   * for other schema type, where the actual schema is resolved from the top level container schema.
-   *
-   * @param name Name of the record.
-   * @return A {@link Schema} of {@link Type#RECORD RECORD} type.
-   */
-  public static Schema recordOf(String name) {
-    Preconditions.checkNotNull(name, "Record name cannot be null.");
-    return new Schema(Type.RECORD, null, null, null, null, name, null, null);
-  }
-
-  /**
-   * Creates a {@link Type#RECORD RECORD} {@link Schema} with the given name and {@link Field Fields}.
-   * The ordering of the fields inside the record would be the same as the one being passed in.
-   *
-   * @param name Name of the record
-   * @param fields All the fields that the record contains.
-   * @return A {@link Schema} of {@link Type#RECORD RECORD} type.
-   */
-  public static Schema recordOf(String name, Field...fields) {
-    return recordOf(name, ImmutableList.copyOf(fields));
-  }
-
-  /**
-   * Creates a {@link Type#RECORD RECORD} {@link Schema} with the given name and {@link Field Fields}.
-   * The ordering of the fields inside the record would be the same as the {@link Iterable#iterator()} order.
-   *
-   * @param name Name of the record
-   * @param fields All the fields that the record contains.
-   * @return A {@link Schema} of {@link Type#RECORD RECORD} type.
-   */
-  public static Schema recordOf(String name, Iterable<Field> fields) {
-    Preconditions.checkNotNull(name, "Record name cannot be null.");
-    ImmutableMap.Builder<String, Field> fieldMapBuilder = ImmutableMap.builder();
-    for (Field field : fields) {
-      fieldMapBuilder.put(field.getName(), field);
-    }
-    Map<String, Field> fieldMap = fieldMapBuilder.build();
-    Preconditions.checkArgument(fieldMap.size() > 0, "No record field provided for %s", name);
-    return new Schema(Type.RECORD, null, null, null, null, name, fieldMap, null);
-  }
-
-  /**
-   * Creates a {@link Type#UNION UNION} {@link Schema} which represents a union of all the given schemas.
-   * The ordering of the schemas inside the union would be the same as the one being passed in.
-   *
-   * @param schemas All the {@link Schema Schemas} constitutes the union.
-   * @return A {@link Schema} of {@link Type#UNION UNION} type.
-   */
-  public static Schema unionOf(Schema...schemas) {
-    return unionOf(ImmutableList.copyOf(schemas));
-  }
-
-  /**
-   * Creates a {@link Type#UNION UNION} {@link Schema} which represents a union of all the given schemas.
-   * The ordering of the schemas inside the union would be the same as the {@link Iterable#iterator()} order.
-   *
-   * @param schemas All the {@link Schema Schemas} constitutes the union.
-   * @return A {@link Schema} of {@link Type#UNION UNION} type.
-   */
-  public static Schema unionOf(Iterable<Schema> schemas) {
-    List<Schema> schemaList = ImmutableList.copyOf(schemas);
-    Preconditions.checkArgument(schemaList.size() > 0, "No union schema provided.");
-    return new Schema(Type.UNION, null, null, null, null, null, null, schemaList);
-  }
-
   private final Type type;
 
   private final Map<String, Integer> enumValues;
@@ -359,21 +125,21 @@ public final class Schema {
 
   private final List<Schema> unionSchemas;
 
-  private String schemaString;
   private SchemaHash schemaHash;
 
-  private Schema(Type type, Set<String> enumValues, Schema componentSchema, Schema keySchema, Schema valueSchema,
-                 String recordName, Map<String, Field> fieldMap, List<Schema> unionSchemas) {
+  public Schema(Type type, Set<String> enumValues, Schema componentSchema, Schema keySchema, Schema valueSchema,
+                String recordName, Map<String, Field> fieldMap, List<Schema> unionSchemas) {
     this.type = type;
     this.enumValues = createIndex(enumValues);
     this.enumIndexes = this.enumValues == null ? null : inverse(this.enumValues);
     this.componentSchema = componentSchema;
     this.keySchema = keySchema;
     this.valueSchema = valueSchema;
-    this.mapSchema = (keySchema == null || valueSchema == null) ? null : Maps.immutableEntry(keySchema, valueSchema);
+    this.mapSchema = (keySchema == null || valueSchema == null) ?
+      null : new AbstractMap.SimpleImmutableEntry<>(keySchema, valueSchema);
     this.recordName = recordName;
     this.fieldMap = populateRecordFields(fieldMap);
-    this.fields = this.fieldMap == null ? null : ImmutableList.copyOf(this.fieldMap.values());
+    this.fields = this.fieldMap == null ? null : new ArrayList<>(this.fieldMap.values());
     this.unionSchemas = unionSchemas;
   }
 
@@ -479,20 +245,6 @@ public final class Schema {
     return (unionSchemas == null || idx < 0 || unionSchemas.size() <= idx) ? null : unionSchemas.get(idx);
   }
 
-  /**
-   * @return a JSON representation of this schema, which can be parsed with {@link #parseJson}
-   */
-  @Override
-  public String toString() {
-    // The follow logic is thread safe, as all the fields buildString() needs are immutable.
-    // It's possible that buildString() get triggered multiple times, but they should yield the same result.
-    String str = schemaString;
-    if (str == null) {
-      schemaString = str = buildString();
-    }
-    return str;
-  }
-
   @Override
   public boolean equals(Object other) {
     if (this == other) {
@@ -535,8 +287,7 @@ public final class Schema {
     if (equals(target)) {
       return true;
     }
-    Multimap<String, String> recordCompared = HashMultimap.create();
-    return checkCompatible(target, recordCompared);
+    return checkCompatible(target, new HashMap<String, String>());
   }
 
   /**
@@ -593,7 +344,7 @@ public final class Schema {
     return firstSchema.getType() == Type.NULL ? unionSchemas.get(1) : firstSchema;
   }
 
-  private boolean checkCompatible(Schema target, Multimap<String, String> recordCompared) {
+  private boolean checkCompatible(Schema target, Map<String, String> recordCompared) {
     if (type.isSimpleType()) {
       if (type == target.getType()) {
         // Same simple type are always compatible
@@ -631,7 +382,7 @@ public final class Schema {
             && valueSchema.checkCompatible(target.valueSchema, recordCompared);
         case RECORD:
           // For every common field (by name), their schema must be compatible
-          if (!recordCompared.containsEntry(recordName, target.recordName)) {
+          if (!(recordCompared.containsKey(recordName) && recordCompared.get(recordName).equals(target.recordName))) {
             recordCompared.put(recordName, target.recordName);
             for (Field field : fields) {
               Field targetField = target.getField(field.getName());
@@ -676,17 +427,17 @@ public final class Schema {
    * @param values Set of values to create index on
    * @return A map from the values to indexes in the set iteration order.
    */
-  private <V> BiMap<V, Integer> createIndex(Set<V> values) {
+  private <V> Map<V, Integer> createIndex(Set<V> values) {
     if (values == null) {
       return null;
     }
 
-    ImmutableBiMap.Builder<V, Integer> builder = ImmutableBiMap.builder();
+    Map<V, Integer> result = new HashMap<>();
     int idx = 0;
     for (V value : values) {
-      builder.put(value, idx++);
+      result.put(value, idx++);
     }
-    return builder.build();
+    return result;
   }
 
   /**
@@ -701,9 +452,9 @@ public final class Schema {
       return null;
     }
 
-    Map<String, Schema> knownRecordSchemas = Maps.newHashMap();
+    Map<String, Schema> knownRecordSchemas = new HashMap<>();
     knownRecordSchemas.put(recordName, this);
-    ImmutableMap.Builder<String, Field> builder = ImmutableMap.builder();
+    Map<String, Field> result = new HashMap<>();
 
     for (Map.Entry<String, Field> fieldEntry : fields.entrySet()) {
       String fieldName = fieldEntry.getKey();
@@ -711,13 +462,13 @@ public final class Schema {
       Schema fieldSchema = resolveSchema(field.getSchema(), knownRecordSchemas);
 
       if (fieldSchema == field.getSchema()) {
-        builder.put(fieldName, field);
+        result.put(fieldName, field);
       } else {
-        builder.put(fieldName, Field.of(fieldName, fieldSchema));
+        result.put(fieldName, Field.of(fieldName, fieldSchema));
       }
     }
 
-    return builder.build();
+    return result;
   }
 
   /**
@@ -734,15 +485,15 @@ public final class Schema {
     switch (schema.getType()) {
       case ARRAY:
         Schema componentSchema = resolveSchema(schema.getComponentSchema(), knownRecordSchemas);
-        return (componentSchema == schema.getComponentSchema()) ? schema : Schema.arrayOf(componentSchema);
+        return (componentSchema == schema.getComponentSchema()) ? schema : Schemas.arrayOf(componentSchema);
       case MAP:
         Map.Entry<Schema, Schema> entry = schema.getMapSchema();
         Schema keySchema = resolveSchema(entry.getKey(), knownRecordSchemas);
         Schema valueSchema = resolveSchema(entry.getValue(), knownRecordSchemas);
         return (keySchema == entry.getKey() && valueSchema == entry.getValue()) ?
-                schema : Schema.mapOf(keySchema, valueSchema);
+                schema : Schemas.mapOf(keySchema, valueSchema);
       case UNION:
-        ImmutableList.Builder<Schema> schemaBuilder = ImmutableList.builder();
+        List<Schema> schemaBuilder = new ArrayList<>();
         boolean changed = false;
         for (Schema input : schema.getUnionSchemas()) {
           Schema output = resolveSchema(input, knownRecordSchemas);
@@ -751,12 +502,14 @@ public final class Schema {
           }
           schemaBuilder.add(output);
         }
-        return changed ? Schema.unionOf(schemaBuilder.build()) : schema;
+        return changed ? Schemas.unionOf(schemaBuilder) : schema;
       case RECORD:
         if (schema.fields == null) {
           // It is a named record that refers to previously defined record
           Schema knownSchema = knownRecordSchemas.get(schema.recordName);
-          Preconditions.checkArgument(knownSchema != null, "Undefined schema %s", schema.recordName);
+          if (knownSchema != null) {
+            throw new IllegalArgumentException(String.format("Undefined schema %s", schema.recordName));
+          }
           return knownSchema;
         } else {
           // It is a concrete schema
@@ -765,27 +518,6 @@ public final class Schema {
         }
     }
     return schema;
-  }
-
-  /**
-   * Helper method to encode this schema into json string.
-   *
-   * @return A json string representing this schema.
-   */
-  private String buildString() {
-    if (type.isSimpleType()) {
-      return '"' + type.name().toLowerCase() + '"';
-    }
-    StringBuilder builder = new StringBuilder();
-    JsonWriter writer = new JsonWriter(CharStreams.asWriter(builder));
-    try {
-      SCHEMA_TYPE_ADAPTER.write(writer, this);
-      writer.close();
-      return builder.toString();
-    } catch (IOException e) {
-      // It should never throw IOException on the StringBuilder Writer, if it does, something very wrong.
-      throw Throwables.propagate(e);
-    }
   }
 
   private <A, B> Map<B, A> inverse(Map<A, B> map) {
